# api-spec/openapi/spec.yaml
openapi: 3.0.3
info:
  title: Book to Course Converter API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /api/v1/pdf/process:
    post:
      summary: Process uploaded PDF
      operationId: processPdf
      description: Accepts a PDF file and returns structure info with chapters and uploadId.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessPdfResponse'
  /api/v1/course/generate:
    post:
      summary: Generate course module from chapter
      operationId: generateCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateCourseRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateCourseResponse'
  /api/v1/books:
    get:
      summary: List books (paginated)
      operationId: getBooksList
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
        - in: query
          name: search
          schema:
            type: string
          required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBooksResponse'
  /api/v1/books/{uploadId}:
    get:
      summary: Get book by uploadId (with table of contents)
      operationId: getBookById
      parameters:
        - in: path
          name: uploadId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDetail'
    delete:
      summary: Delete book by uploadId
      operationId: deleteBook
      parameters:
        - in: path
          name: uploadId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
  /api/v1/tasks/{taskId}/submit:
    post:
      summary: Submit a task answer (non-file)
      operationId: submitTask
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskSubmissionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSubmissionResponse'
  /api/v1/tasks/{taskId}/upload:
    post:
      summary: Submit a task file (upload)
      operationId: submitTaskFile
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSubmissionResponse'
  /api/v1/books/{uploadId}/chapters/{chapterId}/notes:
    get:
      summary: Get generated MDX notes bundle for a chapter
      operationId: getChapterNotes
      parameters:
        - in: path
          name: uploadId
          required: true
          schema:
            type: string
        - in: path
          name: chapterId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Not Found
  /api/v1/books/{uploadId}/chapters/{chapterId}/tasks:
    get:
      summary: Get tasks for a chapter with current user state
      operationId: getChapterTasks
      parameters:
        - in: path
          name: uploadId
          required: true
          schema:
            type: string
        - in: path
          name: chapterId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterTasksResponse'
        '404':
          description: Not Found
  /api/v1/books/{uploadId}/chapters/{chapterId}/images/{filename}:
    get:
      summary: Get generated image asset for a chapter
      operationId: getChapterImage
      parameters:
        - in: path
          name: uploadId
          required: true
          schema:
            type: string
        - in: path
          name: chapterId
          required: true
          schema:
            type: string
        - in: path
          name: filename
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Image bytes
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: Not Found
components:
  schemas:
...
    TaskOption:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
      required: [id, label]
    TaskState:
      type: object
      properties:
        userAnswer:
          type: string
        userAnswers:
          type: array
          items:
            type: string
        userFileName:
          type: string
        evaluation:
          $ref: '#/components/schemas/TaskEvaluation'
        completed:
          type: boolean
      required: [completed]
...
    TaskDefinitionMultipleChoice:
      allOf:
        - $ref: '#/components/schemas/TaskDefinitionBase'
        - type: object
          properties:
            options:
              type: array
              items:
                $ref: '#/components/schemas/TaskOption'
            type:
              type: string
              enum: [multiple-choice]
            correctAnswerId:
              type: string
          required: [type, options]
    TaskDefinitionMultipleSelect:
      allOf:
        - $ref: '#/components/schemas/TaskDefinitionBase'
        - type: object
          properties:
            options:
              type: array
              items:
                $ref: '#/components/schemas/TaskOption'
            type:
              type: string
              enum: [multiple-select]
            correctAnswerIds:
              type: array
              items:
                type: string
          required: [type, options]
...
    TaskDefinition:
      oneOf:
        - $ref: '#/components/schemas/TaskDefinitionMultipleChoice'
        - $ref: '#/components/schemas/TaskDefinitionMultipleSelect'
        - $ref: '#/components/schemas/TaskDefinitionShortAnswer'
        - $ref: '#/components/schemas/TaskDefinitionUploadPdf'
    TaskWithState:
      type: object
      properties:
        definition:
          $ref: '#/components/schemas/TaskDefinition'
        state:
          $ref: '#/components/schemas/TaskState'
      required: [definition, state]
    ChapterTasksResponse:
      type: object
      properties:
        success:
          type: boolean
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskWithState'
      required: [success, tasks]
    Chapter:
      type: object
      required: [chapterId, title, startPage, endPage, isGenerated]
      properties:
        chapterId:
          type: string
        title:
          type: string
        startPage:
          type: integer
          format: int32
        endPage:
          type: integer
          format: int32
        isGenerated:
          type: boolean
        progressData:
          type: object
          properties:
            tasksCount:
              type: integer
            tasksCompleted:
              type: integer
            tasksFailed:
              type: integer
    ProcessPdfData:
      type: object
      properties:
        uploadId:
          type: string
        pageCount:
          type: integer
        wordCount:
          type: integer
        chapters:
          type: array
          items:
            $ref: '#/components/schemas/Chapter'
      required: [uploadId, pageCount, wordCount, chapters]
    ProcessPdfResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/ProcessPdfData'
      required: [success, data]
    GenerateCourseRequest:
      type: object
      properties:
        uploadId:
          type: string
          nullable: false
        chapterId:
          type: string
          nullable: false
      required: [uploadId, chapterId]
    CourseTask:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
...
    GenerateCourseResponse:
      type: object
      properties:
        title:
          type: string
        objectives:
          type: array
          items:
            type: string
        content:
          type: string
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/CourseTask'
      required: [data]
    BookSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        uploadDate:
          type: string
          description: ISO date string (YYYY-MM-DD)
        lastUsedAt:
          type: string
          nullable: true
          description: ISO datetime string
    BookDetail:
      type: object
      required: [id, title, uploadDate, chapters]
      properties:
        id:
          type: string
        title:
          type: string
        uploadDate:
          type: string
        chapters:
          type: array
          items:
            $ref: '#/components/schemas/Chapter'
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, pageSize, total, totalPages]
    PaginatedBooks:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BookSummary'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
        metrics:
          $ref: '#/components/schemas/LibraryMetrics'
      required: [items, pagination, metrics]
    PaginatedBooksResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/PaginatedBooks'
      required: [success, data]
    
    # Aggregated library metrics for the current page and overall dataset
    LibraryMetrics:
      type: object
      properties:
        totalBooks:
          type: integer
          description: Total number of books matching the query (not only current page)
        completedBooks:
          type: integer
          description: Number of books fully completed (all tasks completed and no failures)
        inProgressBooks:
          type: integer
          description: Number of books that have at least one generated course with tasks started or any failures
        failedTasks:
          type: integer
          description: Total count of failed tasks across all matching books
        totalTasks:
          type: integer
          description: Total number of tasks across all matching books
        completedTasks:
          type: integer
          description: Total number of completed tasks across all matching books
        overallProgress:
          type: number
          format: float
          description: Overall progress ratio in range 0..1 across all matching books
      required: [totalBooks, completedBooks, inProgressBooks, failedTasks, totalTasks, completedTasks, overallProgress]
    TaskSubmissionRequest:
      type: object
      properties:
        type:
          type: string
          enum: [multiple-choice, multiple-select, short-answer]
        selectedOptionId:
          type: string
          description: For multiple-choice (option id)
        selectedOptionIds:
          type: array
          items:
            type: string
          description: For multiple-select (option ids)
        textAnswer:
          type: string
          description: For short-answer/code
      required: [type]
    TaskEvaluation:
      type: object
      properties:
        isCorrect:
          type: boolean
        mistakes:
          type: array
          items:
            type: string
        score:
          type: number
    TaskSubmissionResponse:
      type: object
      properties:
        success:
          type: boolean
        evaluation:
          $ref: '#/components/schemas/TaskEvaluation'
        message:
          type: string
      required: [success]